[Feature] As a user, i want to upload multiple photos

+++ src/Command/FixturesCommand.php
@@ -53,6 +53,10 @@ class FixturesCommand extends Command
                 continue;
             }
 
+            if (is_dir($directory . '/' . $file)) {
+                continue;
+            }
+
             unlink($directory . '/' . $file);
         }
     }
